requirements.txt
numpy==1.13.3
scipy==1.0.0b1
tensorflow==1.3.0
Keras==2.0.8
tornado==4.5.2
h5py==2.7.1
sklearn==0.0
Pillow==4.3.0